// src/models/Market-1.model.ts
interface MainCarouselItem {
  URL?: any;
  index: number;
  Order: number;
  title?: string;
  imgUrl?: string;
  BannerFileUrl?: string;
  BannerLargeURL?: string;
  BannerSmallURL?: string;
  imgUrlMobile?: string;
  buttonLik?: string;
  buttonText?: string;
  description?: string;
}

export default MainCarouselItem;